Consolidate operations-related code in src/operation.ts
